Decouple Carousel from Offline module, make common reusable component
