+++ src/common/components/Carousel/carousel.less
@@ -1,0 +1,40 @@
+@import '~common/less/mixins.less';
+@import '~common/less/colors.less';
+@import '~common/less/constants.less';
+
+.container {
+  margin-top: 40px;
+  width: 100%;
+}
+
+.carouselContainer {
+  display: flex;
+  scroll-snap-type: x mandatory;
+  overflow-x: scroll;
+  scrollbar-width: none;
+  &::-webkit-scrollbar {
+    display: none;
+  }
+}
+
+.carousel {
+  display: flex;
+}
+
+.arrowContainer {
+  display: grid;
+  grid-template-columns: repeat(3, 1fr);
+  justify-items: center;
+  margin-bottom: 20px;
+}
+
+.carouselArrow {
+  align-self: center;
+  background: none;
+  font-size: 20px;
+  color: @lightGray;
+  & :hover {
+    color: @gray;
+    cursor: pointer;
+  }
+}
